backend/requirements.txt
Django>=5.0,<7
djangorestframework>=3.15
django-cors-headers>=4.6
python-dotenv>=1.0
djangorestframework-simplejwt>=5.3
Pillow>=10.0
stripe>=11.0,<13
reportlab>=4.0,<5
